cache:replication:e2e: StartStandaloneCacheServer wait for the cache server to be ready
